refactor: share quality helper logic
